# shrimp-farm-web/src/model/farm.ts
/**
 * Shrimp Farm Application
 * Service Contract
 *
 * OpenAPI spec version: 1.0.0
 * Contact: patti_gm2010@live.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */


export interface Farm { 
    id?: string;
    name?: string;
    totalSize?: number;
    nameLocation?: string;
    latitude?: number;
    longitude?: number;
    idPonds?: Array<string>;
}